refactor: separate music upload modal
